Implemented add_message endpoint and cleaned up in the code
